Make rudimentary timeslots list page
