+++ src/argus_htmx/timeslots/urls.py
@@ -1,21 +1,9 @@
-from django.http import HttpResponse
-from django.template import Template, RequestContext
 from django.urls import path
 
-
-def placeholder(request):
-    template = Template(
-        """{% extends "htmx/base.html" %}
-        {% block main %}
-        <h1>TIMESLOT PLACEHOLDER</h1>
-        {% endblock main %}
-        """
-    )
-    context = RequestContext(request)
-    return HttpResponse(template.render(context))
+from .views import TimeslotListView
 
 
 app_name = "htmx"
 urlpatterns = [
-    path("", placeholder, name="timeslot-placeholder"),
+    path("", TimeslotListView.as_view(), name="timeslot-list"),
 ]
